+++ src/main/java/racingcar/models/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.models;
 
 @FunctionalInterface
 public interface NumberGenerator {
