+++ src/main/java/racingcar/views/StandardConsole.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.views;
 
 import java.util.Scanner;
