+++ src/test/java/racingcar/RacingGameTest.java
@@ -5,8 +5,9 @@ import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import racingcar.engine.domain.Car;
-import racingcar.engine.domain.RacingGame;
+import racingcar.models.Car;
+import racingcar.models.RacingGame;
+import racingcar.models.RandomNumberGenerator;
 
 public class RacingGameTest {
